% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runflexModels.R
\name{runflexModels}
\alias{runflexModels}
\title{A function to run the models of flexmirt
This function is just a demo.}
\usage{
runflexModels(Path.file, Path.flex = getwd())
}
\description{
A function to run the models of flexmirt
This function is just a demo.
}

